Fix updating status by preloading jobs
The status presenter methods to queries on jobs, but
all those queries include .count or .any? etc., queries
that Rails avoids loading the actual Job models for. Which means
after_find on the Job models is never triggered. Which means the cached
status value for the Job records is never updated.

I preload the jobs in the index controller, which avoids this problem.

In the future, we should look for a cleaner solution.