unroll compact code for readability
I prefer the original two lines where we create the
Job objects, setup the dependency, and append to the Jobs array.

When you know what you are doing it is simple and straight forward.

However, since this is serving as an example for other developers,
unrolling the 2 lines into several more, where we create the Job
objects in one line, setup the dependency in another, and add them
to the array in the final lines, with comments, seems to make the code
more readable and clearer on exactly what is happening.

If I'm wrong, and the pithy 2 line approach is just as readable, I'll
gladly revert to that.